requirements.txt
Pillow
qrcode
bitarray
pypng